split out VGuidelinesTextItem class into its own file
